ebuild push, mainly libvirt related
- app-emulation/libvirt-glib-3.0.0
0 tests failed!

- app-emulation/libvirt-5.9.0
- app-emulation/libvirt-python-5.9.0
Didn't test it but...it works! Lol.
To note: Additional patches have been provided for libvirt-5.9.0.
	- 5.9.0-0005-release-unused-memory-back-to-the-system.patch -- self-explanatory
	- 5.9.0-Replaced-deprecated-var-run-links-to-run.patch -- Replaced all /var/run links to /run to get rid of the annoying deprecated warning

- app-emulation/looking-glass-9999
Works, difficult to set up. You must look for an obscure signed IVSHMEM .inf for Windows, I can't remember where I found it. Instructions on signing your own drivers (requiring +5GB of dev software) is outdated for Windows 1909
I had it working for a bit, but it was buggy due to running it off an RX580 (reset bug?)